+++ helper.sh
@@ -9,6 +9,7 @@
 
 DIR="${BASH_SOURCE%/*}"
 if [ ! -d "$DIR" ]; then DIR="$PWD"; fi
+logFile=$DIR/log/day_$(date '+%Y-%m-%d').log
 . "$DIR/app/lib.sh"
 . "$DIR/app/cmd.sh"
 
@@ -33,12 +34,12 @@ boot()
     for conf in $(find "$DIR/config" -name '*.yml' -or -name '*.yaml')
     do
         # read .yml file
-        eval $(parse_yaml ${conf} "_")
+        eval "$(parse_yaml ${conf} '_')"
     done
 
+    subCmd $@
     draw _ 24 'menu'
     helperAlias
-    subCmd $@
     menu "option" $1
 }