+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 CHANGELOG
 =========
 
+1.3.0
+-----
+
+ * refactor mysql cli trough mysql_config_editor
+ * mySQL show users
+ * mySQL show user privileges
+ * mySQL DB size
+ * run backup without `sudo bash helper.sh`
+ * add [PSBook](https://github.com/svnvcristea/PSBook) to xbuild
+ * SugarDevEnv install and run phpunit
+ * day on | off
+ * check and set MySQL root pass while xbuild
+ * mySQL fix for SuSe and Mint to use mysql_config_editor
+
 1.2.0
 -----
+
  * git clone GitHub repo
  * VPN status
  * use encoded sudo password